completed lab13 (try to fix Radix Sort)

+++ lab13/RadixSort.java
@@ -26,13 +26,13 @@ public class RadixSort {
         for (int d = max - 1; d >= 0; d--) {
             int[] count = new int[R];
             for (int i = 0; i < N; i++) {
-                count[(int) asciis[i].charAt(d)]++;
+                count[asciis[i].charAt(d)]++;
             }
             for (int r = 1; r < R; r++) {
                 count[r] += count[r - 1];
             }
             for (int i = 0; i < N; i++) {
-                aux[count[(int) asciis[i].charAt(d)]++] = asciis[i];
+                aux[count[asciis[i].charAt(d)]++] = asciis[i];
             }
             for (int i = 0; i < N; i++) {
                 asciis[i] = aux[i];
